refactor: command 폴더 생성
