+++ src/main/java/blackjack/BlackjackGame.java
@@ -5,12 +5,11 @@ import static blackjack.util.ExceptionHandler.retryUntilSuccess;
 import blackjack.model.bet.BetAmount;
 import blackjack.model.bet.BetAmounts;
 import blackjack.model.card.CardProvider;
-import blackjack.model.card.HitCommand;
+import blackjack.model.command.HitCommand;
 import blackjack.model.gameresult.ProfitResult;
 import blackjack.model.user.User;
 import blackjack.model.user.Users;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.Function;
