+++ src/main/java/blackjack/model/command/HitCommand.java
@@ -1,4 +1,4 @@
-package blackjack.model.card;
+package blackjack.model.command;
 
 import java.util.regex.Pattern;
