+++ src/test/java/blackjack/model/card/HitCommandTest.java
@@ -1,10 +1,11 @@
 package blackjack.model.card;
 
-import static blackjack.model.card.HitCommand.*;
+import static blackjack.model.command.HitCommand.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import blackjack.model.command.HitCommand;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
